app/services/binance_service.py
import binance
from app.extensions import db
from app.models.wallet import Wallet

class BinanceService:
    def __init__(self):
        self.client = binance.Client(api_key='your_binance_api_key', api_secret='your_binance_api_secret')

    def create_wallet(self):
        # Implement wallet creation logic here
        pass

    def get_total_balance(self, user_id):
        wallets = Wallet.query.filter_by(user_id=user_id).all()
        total_balance = 0
        for wallet in wallets:
            balance = self.get_wallet_balance(wallet.id)
            total_balance += balance
        return total_balance

    def get_wallet_balance(self, wallet_id):
        # Implement wallet balance retrieval logic here
        pass

    def send_assets(self, from_wallet, to_wallet, amount):
        # Implement asset sending logic here
        pass

    def get_transactions_history(self, user_id):
        # Implement transaction history retrieval logic here
        pass
