requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
requests==2.27.1
bcrypt==3.2.0
Flask-OAuthlib==0.9.5
cryptography==3.4.7
Flask-Limiter==1.4
gunicorn
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.3.1
python-binance==1.0.12